Added statistics methode
